Don't put name into it's own folder.
